Check if file exists

+++ generate_embeddings.py
@@ -2,6 +2,7 @@ from datasets import load_dataset
 from sentence_transformers import SentenceTransformer
 import pickle
 import tqdm
+import os
 
 # Define the custom download directory
 custom_path = "/globalscratch/ucl/cental/troux/corpus/stsb"
@@ -16,13 +17,20 @@ dataset = load_dataset("stsb_multi_mt", name=lang, cache_dir=custom_path)
 
 model = SentenceTransformer(model_name)
 
-id2embedding = {}
-# for id, example in enumerate(dataset["train"]):
-for id, example in tqdm.tqdm(enumerate(dataset["train"]), total=len(dataset["train"])):
-    embedding = model.encode(example[sentence_key]) # sentence1 or sentence2
-    id2embedding[id] = embedding
+# Check if the embeddings are already generated
+if os.path.exists(saved_embeddings_path):
+    print(f"Embeddings already exist at {saved_embeddings_path}. Loading...")
+    with open(saved_embeddings_path, "rb") as f:
+        id2embedding = pickle.load(f)
+    print(f"Loaded {len(id2embedding)} embeddings")
+else:
+    id2embedding = {}
+    # for id, example in enumerate(dataset["train"]):
+    for id, example in tqdm.tqdm(enumerate(dataset["train"]), total=len(dataset["train"])):
+        embedding = model.encode(example[sentence_key]) # sentence1 or sentence2
+        id2embedding[id] = embedding
 
-# Save the embeddings to a file
-with open(saved_embeddings_path, "wb") as f:
-    pickle.dump(id2embedding, f)
-print(f"Embeddings saved to {saved_embeddings_path}")
+    # Save the embeddings to a file
+    with open(saved_embeddings_path, "wb") as f:
+        pickle.dump(id2embedding, f)
+    print(f"Embeddings saved to {saved_embeddings_path}")
